Dynamic thread pool monitoring module refactoring
