+++ hippo4j-spring-boot/hippo4j-core-spring-boot-starter/pom.xml
@@ -74,19 +74,7 @@
 
         <dependency>
             <groupId>cn.hippo4j</groupId>
-            <artifactId>hippo4j-monitor-log</artifactId>
-            <version>${revision}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>cn.hippo4j</groupId>
-            <artifactId>hippo4j-monitor-prometheus</artifactId>
-            <version>${revision}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>cn.hippo4j</groupId>
-            <artifactId>hippo4j-monitor-es</artifactId>
+            <artifactId>hippo4j-monitor-base</artifactId>
             <version>${revision}</version>
         </dependency>
     </dependencies>
