+++ hippo4j-spring-boot/hippo4j-spring-boot-starter-monitor/pom.xml
@@ -4,16 +4,15 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>cn.hippo4j</groupId>
-        <artifactId>hippo4j-all</artifactId>
+        <artifactId>hippo4j-spring-boot</artifactId>
         <version>${revision}</version>
     </parent>
-    <artifactId>hippo4j-spring-boot</artifactId>
+    <artifactId>hippo4j-spring-boot-starter-monitor</artifactId>
     <packaging>pom</packaging>
 
     <modules>
-        <module>hippo4j-core-spring-boot-starter</module>
-        <module>hippo4j-spring-boot-starter</module>
-        <module>hippo4j-spring-boot-starter-adapter</module>
-        <module>hippo4j-spring-boot-starter-monitor</module>
+        <module>hippo4j-spring-boot-starter-monitor-es</module>
+        <module>hippo4j-spring-boot-starter-monitor-log</module>
+        <module>hippo4j-spring-boot-starter-monitor-prometheus</module>
     </modules>
 </project>
